refactor: create new high level chunking strategies folder to contain layout and line-sentence
