+++ src/ingestion_pipeline/chunking/document_chunker_factory.py
@@ -3,14 +3,14 @@
 import logging
 
 from ingestion_pipeline.chunking.base_document_chunker import DocumentChunker
-from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout.types.key_value.layout_key_value import KeyValueChunker
-from ingestion_pipeline.chunking.layout.types.layout_text import LayoutTextChunkingStrategy
-from ingestion_pipeline.chunking.layout.types.list.list_chunker import LayoutListChunkingStrategy
-from ingestion_pipeline.chunking.layout.types.table.layout_table import LayoutTableChunkingStrategy
 from ingestion_pipeline.chunking.line_sentence_chunker.config import LineSentenceChunkingConfig
 from ingestion_pipeline.chunking.line_sentence_splitter import LineBasedDocumentChunker
-from ingestion_pipeline.chunking.textract_layout_chunker import TextractLayoutDocumentChunker
+from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.textract_layout_chunker import TextractLayoutDocumentChunker
+from ingestion_pipeline.chunking.strategies.layout.types.key_value.layout_key_value import KeyValueChunker
+from ingestion_pipeline.chunking.strategies.layout.types.layout_text import LayoutTextChunkingStrategy
+from ingestion_pipeline.chunking.strategies.layout.types.list.list_chunker import LayoutListChunkingStrategy
+from ingestion_pipeline.chunking.strategies.layout.types.table.layout_table import LayoutTableChunkingStrategy
 from ingestion_pipeline.config import settings
 
 ALLOWED_CHUNKER_TYPES = {"layout", "linear-sentence-splitter"}
