+++ src/ingestion_pipeline/chunking/strategies/layout/textract_layout_chunker.py
@@ -11,12 +11,11 @@ from textractor.entities.document import Document
 
 from ingestion_pipeline.chunking.base_document_chunker import ChunkError, DocumentChunker
 from ingestion_pipeline.chunking.exceptions import ChunkException
-from ingestion_pipeline.chunking.layout.types.merge.chunk_merger import ChunkMerger
+from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata, ProcessedDocument
+from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.types.base import ChunkingStrategyHandler
+from ingestion_pipeline.chunking.strategies.layout.types.merge.chunk_merger import ChunkMerger
 from ingestion_pipeline.chunking.verbose_page_debug_logger import is_verbose_page_debug, log_verbose_page_debug
-
-from .layout.layout_chunking_config import LayoutChunkingConfig
-from .layout.types.base import ChunkingStrategyHandler
-from .schemas import DocumentChunk, DocumentMetadata, ProcessedDocument
 
 logger = logging.getLogger(__name__)
