+++ src/ingestion_pipeline/chunking/strategies/layout/types/base.py
@@ -3,8 +3,10 @@
 from abc import ABC, abstractmethod
 from typing import List, Optional
 
-from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig  # ← Changed import
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
+from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import (
+    LayoutChunkingConfig,
+)  # ← Changed import
 
 
 class ChunkingStrategyHandler(ABC):
