+++ src/ingestion_pipeline/chunking/strategies/layout/types/table/base.py
@@ -7,8 +7,8 @@ from typing import List, Optional
 from textractor.entities.bbox import BoundingBox
 from textractor.entities.layout import Layout
 
-from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
 from ingestion_pipeline.chunking.schemas import DocumentChunk, DocumentMetadata
+from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
 
 logger = logging.getLogger(__name__)
