+++ tests/chunking/strategies/test_layout_text_strategy.py
@@ -4,9 +4,9 @@ from unittest.mock import MagicMock, call
 
 import pytest
 
-import ingestion_pipeline.chunking.layout.types.layout_text as layout_text_module
-from ingestion_pipeline.chunking.layout.layout_chunking_config import LayoutChunkingConfig
-from ingestion_pipeline.chunking.layout.types.layout_text import LayoutTextChunkingStrategy
+import ingestion_pipeline.chunking.strategies.layout.types.layout_text as layout_text_module
+from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.strategies.layout.types.layout_text import LayoutTextChunkingStrategy
 
 
 @pytest.fixture
